dom/tests/mochitest/webapps/test_install_receipts.xul

<?xml version="1.0"?>

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->

<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Mozilla Bug 741549">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="head.js"/>
  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=741549"
     target="_blank">Mozilla Bug 741549</a>
  </body>

<script>

SimpleTest.waitForExplicitFinish();

var url = "http://test/chrome/dom/tests/mochitest/webapps/apps/basic.webapp";

var receipts = [
  "a",
  0,
  true,
  false,
  null,
  undefined,
  { prop: 1 },
  { nest: { ed: "hi" } },
  NaN,
  Infinity,
];

confirmNextInstall();
navigator.mozApps.install(url, { receipts: receipts }).onsuccess =
function onInstall() {
  is(this.result.receipts[0], "a", "receipts[0]");
  is(this.result.receipts[1], 0, "receipts[1]");
  is(this.result.receipts[2], true, "receipts[2]");
  is(this.result.receipts[3], false, "receipts[3]");
  is(this.result.receipts[4], null, "receipts[4]");
  is(this.result.receipts[5], undefined, "receipts[5]");
  is(this.result.receipts[6].prop, 1, "receipts[6].prop");
  is(this.result.receipts[7].nest.ed, "hi", "receipts[7].nest.ed");
  todo_is(this.result.receipts[8], NaN, "receipts[8]");
  todo_is(this.result.receipts[9], Infinity, "receipts[9]");

  navigator.mozApps.mgmt.uninstall(this.result).onsuccess = function onUninstall() {
    SimpleTest.finish();
  }
};

</script>
</window>
